user-interface and final def function
I googled for user interface to solve to problem of how for the user to
enter a "family member expression". I found GUI on python-course.eu,
used the code, and tried to figure out what code belongs to which
graphic and then adjusted it to my program.
Afterwards, I defined the langvisual function to put all the previously
defined functions together. Most work is done, but: I need to build more
stuff around the program so the 500 lines will be achieved.