window titles, solved: if one expression is part of two languages, delay
I founf some more funny functions belonging to the turtle module. One of
them is to add a title to the uppopping turtle screen.
I solved the problem that accurs if one family member expression is part
of two languages. First: I got rid of the exitonclick-unction at the
very end of each flag function. Second: I imported the time module in
order to delay the appearance of the next language flag so the user will
have some seconds to check which flags just popped up.