def falseentry function, problem
In case the user enters a word that is not in the lists of epxressions
for family members, the programm is supposed to return a red screen
saying "false entry". Unfortunately, I still need to figure out how to
include to the script without popping up ANYTIME a words is entered. It
is only supposed to open if something wrong is typed in.