problem solved, lines shorter
I made some lines shorter because according to the python style book is
that better.
Besides, I found the reason why my structure before would not work:
compairing the element of insert_word with the list langlist does not
work because it cannot be looked into the nested lists. So i added
markes in order to check if the if block was executed or not.